backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy==2.0.30
httpx==0.27.0
python-dotenv==1.0.1
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
bcrypt
